Respect conduit template file convention
